Fix/std max compiling error

See merge request desktop/desktop!1471

+++ src/MEGASync/transfers/gui/DuplicatedNodeDialogs/DuplicatedNodeItem.cpp
@@ -199,8 +199,8 @@ QDateTime DuplicatedRemoteItem::getModifiedTime()
 
 qint64 DuplicatedRemoteItem::getNodeSize()
 {
-    auto nodeSize = mInfo->getRemoteConflictNode()->getSize();
-    return std::max(nodeSize, 0L);
+    auto nodeSize = static_cast<qint64>(mInfo->getRemoteConflictNode()->getSize());
+    return std::max(nodeSize, 0LL);
 }
 
 bool DuplicatedRemoteItem::isFile() const
